feat(rightdown): remove v1 support and migration references
- Simplified types.ts to only support v2 config format
- Removed all v1 validation logic from config-reader
- Updated all test files to remove v1/v2 distinctions
- Renamed test fixtures for clarity
- Updated package.json description
- Updated CLAUDE.md documentation for v2-only architecture
- Renamed plan document to remove version number
- This is a clean break with no backward compatibility
