+++ packages/rightdown/package.json
@@ -1,7 +1,7 @@
 {
   "name": "@outfitter/rightdown",
   "version": "2.0.0",
-  "description": "Opinionated Markdown linting and formatting for correct documentation",
+  "description": "Unified Markdown formatter orchestrating code block formatting tools",
   "keywords": [
     "markdown",
     "linting",
@@ -52,11 +52,13 @@
     }
   },
   "devDependencies": {
+    "@biomejs/biome": "^2.0.0",
     "@outfitter/typescript-config": "workspace:*",
     "@types/js-yaml": "^4.0.9",
     "@types/node": "^22.7.4",
     "@types/yargs": "^17.0.33",
     "markdownlint": "^0.34.0",
+    "prettier": "^3.5.3",
     "tsup": "^8.3.0",
     "typescript": "^5.6.3",
     "vitest": "^2.1.5"
