+++ packages/rightdown/src/__tests__/fixtures/configs/basic.yaml
