+++ packages/rightdown/src/__tests__/fixtures/configs/invalid-yaml.yaml
@@ -2,4 +2,5 @@
 invalid:
   - missing closing bracket
   [ item1
-  item2:
+item2:
+  nested: value
