--- packages/rightdown/src/__tests__/fixtures/configs/v1-legacy.yaml
@@ -1,24 +1,0 @@
-# Legacy v1 config (no version field)
-# This should continue to work with existing behavior
-
-# Built-in preset
-preset: standard
-
-# Custom rule overrides
-rules:
-  line-length: false
-  list-marker-space: false
-  no-duplicate-heading: false
-
-# Ignore patterns
-ignores:
-  - "node_modules/**"
-  - "dist/**"
-  - "build/**"
-
-# Custom terminology
-terminology:
-  - incorrect: "javascript"
-    correct: "JavaScript"
-  - incorrect: "Github"
-    correct: "GitHub"
