+++ packages/rightdown/src/__tests__/formatters/biome.test.ts
@@ -40,7 +40,7 @@ describe('BiomeFormatter', () => {
   describe('isAvailable', () => {
     it('should return true when @biomejs/biome is installed', async () => {
       const result = await formatter.isAvailable();
-      
+
       expect(isSuccess(result)).toBe(true);
       if (result.success) {
         expect(result.data).toBe(true);
@@ -49,7 +49,7 @@ describe('BiomeFormatter', () => {
 
     it('should return false when @biomejs/biome is not installed', async () => {
       const testFormatter = new BiomeFormatter();
-      
+
       vi.doMock('@biomejs/biome', () => {
         const error = new Error('Cannot find module');
         (error as any).code = 'MODULE_NOT_FOUND';
@@ -57,12 +57,12 @@ describe('BiomeFormatter', () => {
       });
 
       const result = await testFormatter.isAvailable();
-      
+
       expect(isSuccess(result)).toBe(true);
       if (result.success) {
         expect(result.data).toBe(false);
       }
-      
+
       vi.doUnmock('@biomejs/biome');
     });
   });
@@ -71,11 +71,11 @@ describe('BiomeFormatter', () => {
     it('should return biome version', async () => {
       // Mock the package.json import
       vi.doMock('@biomejs/biome/package.json', () => ({
-        default: { version: '1.5.0' }
+        default: { version: '1.5.0' },
       }));
-      
+
       const result = await formatter.getVersion();
-      
+
       expect(isSuccess(result)).toBe(true);
       if (result.success) {
         expect(result.data).toBe('1.5.0');
@@ -84,7 +84,7 @@ describe('BiomeFormatter', () => {
 
     it('should handle missing biome gracefully', async () => {
       const testFormatter = new BiomeFormatter();
-      
+
       vi.doMock('@biomejs/biome', () => {
         const error = new Error('Cannot find module');
         (error as any).code = 'MODULE_NOT_FOUND';
@@ -92,12 +92,12 @@ describe('BiomeFormatter', () => {
       });
 
       const result = await testFormatter.getVersion();
-      
+
       expect(isFailure(result)).toBe(true);
       if (!result.success) {
         expect(result.error.code).toBe('NOT_FOUND');
       }
-      
+
       vi.doUnmock('@biomejs/biome');
     });
   });
@@ -106,7 +106,7 @@ describe('BiomeFormatter', () => {
     it('should format JavaScript code', async () => {
       const code = 'const x=1;const y=2;';
       const result = await formatter.format(code, 'javascript');
-      
+
       expect(isSuccess(result)).toBe(true);
       if (result.success) {
         expect(result.data).toBe('const x = 1;\nconst y = 2;\n');
@@ -116,7 +116,7 @@ describe('BiomeFormatter', () => {
     it('should format TypeScript code', async () => {
       const code = 'interface User{name:string;age:number}';
       const result = await formatter.format(code, 'typescript');
-      
+
       expect(isSuccess(result)).toBe(true);
       if (result.success) {
         expect(result.data).toContain('interface User {\n');
@@ -128,7 +128,7 @@ describe('BiomeFormatter', () => {
     it('should format JSON', async () => {
       const code = '{"name":"test","value":123}';
       const result = await formatter.format(code, 'json');
-      
+
       expect(isSuccess(result)).toBe(true);
       if (result.success) {
         expect(result.data).toBe('{\n\t"name": "test",\n\t"value": 123\n}\n');
@@ -142,9 +142,9 @@ describe('BiomeFormatter', () => {
         indentWidth: 2,
         semicolons: 'asNeeded',
       };
-      
+
       const result = await formatter.format(code, 'javascript', options);
-      
+
       expect(isSuccess(result)).toBe(true);
       if (result.success) {
         expect(result.data).toBe('const x = 1\n'); // No semicolon with asNeeded
@@ -154,7 +154,7 @@ describe('BiomeFormatter', () => {
     it('should handle syntax errors', async () => {
       const code = 'const x = {';
       const result = await formatter.format(code, 'javascript');
-      
+
       expect(isFailure(result)).toBe(true);
       if (!result.success) {
         expect(result.error.code).toBe('INTERNAL_ERROR');
@@ -165,7 +165,7 @@ describe('BiomeFormatter', () => {
     it('should handle unsupported languages', async () => {
       const code = '.container { display: flex; }';
       const result = await formatter.format(code, 'css');
-      
+
       expect(isFailure(result)).toBe(true);
       if (!result.success) {
         expect(result.error.code).toBe('INTERNAL_ERROR');
@@ -176,7 +176,7 @@ describe('BiomeFormatter', () => {
     it('should format JSX code', async () => {
       const code = 'const App=()=><div>Hello</div>;';
       const result = await formatter.format(code, 'jsx');
-      
+
       expect(isSuccess(result)).toBe(true);
       if (result.success) {
         expect(result.data).toContain('const App = () => <div>Hello</div>;');
@@ -186,7 +186,7 @@ describe('BiomeFormatter', () => {
     it('should format TSX code', async () => {
       const code = 'const App:React.FC=()=><div>Hello</div>;';
       const result = await formatter.format(code, 'tsx');
-      
+
       expect(isSuccess(result)).toBe(true);
       if (result.success) {
         expect(result.data).toContain('const App: React.FC = () => <div>Hello</div>;');
@@ -197,7 +197,7 @@ describe('BiomeFormatter', () => {
   describe('getSupportedLanguages', () => {
     it('should return list of supported languages', () => {
       const languages = formatter.getSupportedLanguages();
-      
+
       expect(languages).toContain('javascript');
       expect(languages).toContain('typescript');
       expect(languages).toContain('jsx');
@@ -215,26 +215,24 @@ describe('BiomeFormatter', () => {
       await formatter.format('const x = 1;', 'javascript');
       await formatter.format('const y = 2;', 'javascript');
       await formatter.format('const z = 3;', 'javascript');
-      
+
       // Biome.create should only be called once
       const biomeMock = await vi.importMock('@biomejs/biome');
       expect(biomeMock.Biome.create).toHaveBeenCalledTimes(1);
     });
 
     it('should handle large files efficiently', async () => {
-      const largeCode = Array(1000)
-        .fill('const x = 1;')
-        .join('\n');
-      
+      const largeCode = Array(1000).fill('const x = 1;').join('\n');
+
       mockFormatContent.mockResolvedValueOnce({
         content: largeCode, // Just return the same for the test
         diagnostics: [],
       });
-      
+
       const start = Date.now();
       const result = await formatter.format(largeCode, 'javascript');
       const duration = Date.now() - start;
-      
+
       expect(isSuccess(result)).toBe(true);
       expect(duration).toBeLessThan(1000); // Should complete within 1 second
     });
