+++ packages/rightdown/src/cli.ts
@@ -4,28 +4,31 @@ import yargs from 'yargs';
 import { hideBin } from 'yargs/helpers';
 
 // Import command handlers
-import { initCommand } from './commands/init.js';
 import { formatCommand } from './commands/format.js';
-import { rulesListCommand, rulesUpdateCommand, rulesForgetCommand } from './commands/rules.js';
-import { configPresetCommand, configIgnoreCommand } from './commands/config.js';
-import { lintCommand } from './commands/lint.js';
+import { initCommand } from './commands/init.js';
 
 // Build CLI
 yargs(hideBin(process.argv))
-  // Default command: lint files
+  // Default command: format files
   .command(
     '$0 [files..]',
-    'Lint Markdown files',
+    'Format markdown files with code blocks',
     (yargs) => {
       return yargs
         .positional('files', {
-          describe: 'Files to lint',
+          describe: 'Files to format',
           type: 'string',
           array: true,
-          default: ['.'],
+          default: ['**/*.md'],
         })
-        .option('fix', {
-          describe: 'Fix issues automatically',
+        .option('write', {
+          alias: 'w',
+          describe: 'Write formatted output back to files',
+          type: 'boolean',
+          default: false,
+        })
+        .option('check', {
+          describe: 'Check if files are formatted correctly',
           type: 'boolean',
           default: false,
         })
@@ -33,9 +36,11 @@ yargs(hideBin(process.argv))
           alias: 'c',
           describe: 'Configuration file path',
           type: 'string',
-        });
+          default: '.rightdown.config.yaml',
+        })
+        .conflicts('write', 'check');
     },
-    lintCommand,
+    formatCommand,
   )
 
   // Init command
@@ -43,146 +48,22 @@ yargs(hideBin(process.argv))
     'init [preset]',
     'Initialize Rightdown configuration',
     (yargs) => {
-      return yargs.positional('preset', {
-        describe: 'Preset to use',
-        type: 'string',
-        choices: ['strict', 'standard', 'relaxed'],
-      });
+      return yargs
+        .positional('preset', {
+          describe: 'Preset to use',
+          type: 'string',
+          choices: ['strict', 'standard', 'relaxed'],
+          default: 'standard',
+        })
+        .option('force', {
+          alias: 'f',
+          describe: 'Overwrite existing configuration',
+          type: 'boolean',
+          default: false,
+        });
     },
     initCommand,
   )
-
-  // Format command
-  .command(
-    'format [source] [path]',
-    'Format Markdown from various sources',
-    (yargs) => {
-      return yargs
-        .positional('source', {
-          describe: 'Source to format from',
-          type: 'string',
-          choices: ['file'],
-        })
-        .positional('path', {
-          describe: 'File path (when source is "file")',
-          type: 'string',
-        })
-        .option('input', {
-          describe: 'Read from stdin',
-          type: 'boolean',
-        })
-        .option('text', {
-          describe: 'Format inline text',
-          type: 'string',
-        })
-        .check((argv) => {
-          // Ensure at least one source is specified
-          if (!argv.source && !argv.input && !argv.text) {
-            throw new Error('Must specify a source: file, --input, or --text');
-          }
-          if (argv.source === 'file' && !argv.path) {
-            throw new Error('File path required when source is "file"');
-          }
-          return true;
-        })
-        .option('preset', {
-          alias: 'p',
-          describe: 'Preset to use for formatting',
-          type: 'string',
-          choices: ['strict', 'standard', 'relaxed'],
-        })
-        .option('output', {
-          describe: 'Output to stdout',
-          type: 'boolean',
-          default: true,
-        });
-    },
-    formatCommand,
-  )
-
-  // Rules commands
-  .command('rules', 'Manage Markdown rules', (yargs) => {
-    return yargs
-      .command(
-        'list [preset]',
-        'List rules',
-        (yargs) => {
-          return yargs.positional('preset', {
-            describe: 'Show rules for specific preset',
-            type: 'string',
-            choices: ['strict', 'standard', 'relaxed'],
-          });
-        },
-        rulesListCommand,
-      )
-      .command(
-        'update <rules..>',
-        'Update rule configurations',
-        (yargs) => {
-          return yargs
-            .positional('rules', {
-              describe: 'Rules to update (e.g., md003 atx md013 --line-length 80)',
-              type: 'string',
-              array: true,
-              demandOption: true,
-            })
-            .strict(false); // Allow dynamic options for rules
-        },
-        rulesUpdateCommand,
-      )
-      .command(
-        'forget <rules..>',
-        'Remove rule overrides',
-        (yargs) => {
-          return yargs.positional('rules', {
-            describe: 'Rules to forget',
-            type: 'string',
-            array: true,
-            demandOption: true,
-          });
-        },
-        rulesForgetCommand,
-      )
-      .demandCommand(1, 'You need to specify a rules subcommand');
-  })
-
-  // Config commands
-  .command('config', 'Manage configuration', (yargs) => {
-    return yargs
-      .command(
-        'preset <name>',
-        'Set configuration preset',
-        (yargs) => {
-          return yargs.positional('name', {
-            describe: 'Preset name',
-            type: 'string',
-            choices: ['strict', 'standard', 'relaxed'],
-            demandOption: true,
-          });
-        },
-        configPresetCommand,
-      )
-      .command(
-        'ignore <patterns..>',
-        'Manage ignore patterns',
-        (yargs) => {
-          return yargs
-            .positional('patterns', {
-              describe: 'Patterns to add or remove',
-              type: 'string',
-              array: true,
-              demandOption: true,
-            })
-            .option('remove', {
-              describe: 'Remove patterns instead of adding',
-              type: 'boolean',
-              default: false,
-            });
-        },
-        configIgnoreCommand,
-      )
-      .demandCommand(1, 'You need to specify a config subcommand');
-  })
 
   // Global options
   .option('quiet', {
@@ -202,18 +83,19 @@ yargs(hideBin(process.argv))
   .help()
   .alias('help', 'h')
   .version()
-  .alias('version', 'v')
+  .alias('version', 'V')
   .strict()
   .recommendCommands()
   .showHelpOnFail(false, 'Use --help for available options')
 
   // Examples
-  .example('$0', 'Lint all Markdown files in current directory')
-  .example('$0 --fix README.md', 'Fix issues in README.md')
-  .example('$0 init standard', 'Initialize with standard preset')
-  .example('$0 format --text "# My Title" --preset strict', 'Format text with strict preset')
-  .example('$0 rules update md013 --line-length 100', 'Update line length rule')
-  .example('$0 rules forget md050', 'Remove emphasis style override')
+  .example('$0 init', 'Initialize with standard preset')
+  .example('$0 init strict', 'Initialize with strict preset')
+  .example('$0', 'Format all markdown files (output to stdout)')
+  .example('$0 --write', 'Format all markdown files in place')
+  .example('$0 --check', 'Check if files are formatted correctly')
+  .example('$0 README.md --write', 'Format specific file')
+  .example('$0 docs/**/*.md --write', 'Format all docs')
 
   .epilogue('For more information, visit https://github.com/outfitter-dev/rightdown')
   .parse();
