+++ packages/rightdown/src/core/errors.ts
@@ -9,19 +9,19 @@ export const RIGHTDOWN_ERROR_CODES = {
   INVALID_YAML: ErrorCode.VALIDATION_ERROR,
   INVALID_CONFIG: ErrorCode.VALIDATION_ERROR,
   UNSUPPORTED_VERSION: ErrorCode.VALIDATION_ERROR,
-  
+
   // Formatter errors
   FORMATTER_NOT_FOUND: ErrorCode.NOT_FOUND,
   FORMATTER_FAILED: ErrorCode.INTERNAL_ERROR,
   FORMATTER_TIMEOUT: ErrorCode.INTERNAL_ERROR,
-  
+
   // AST processing errors
   PARSE_ERROR: ErrorCode.VALIDATION_ERROR,
   INVALID_MARKDOWN: ErrorCode.VALIDATION_ERROR,
-  
+
   // General errors
   VALIDATION_ERROR: ErrorCode.VALIDATION_ERROR,
   IO_ERROR: ErrorCode.INTERNAL_ERROR,
 } as const;
 
-export type RightdownErrorCode = typeof RIGHTDOWN_ERROR_CODES[keyof typeof RIGHTDOWN_ERROR_CODES];
+export type RightdownErrorCode = (typeof RIGHTDOWN_ERROR_CODES)[keyof typeof RIGHTDOWN_ERROR_CODES];
