+++ packages/rightdown/src/core/presets.ts
@@ -17,7 +17,7 @@ export const PRESETS: Record<string, PresetConfig> = {
       'heading-style': { style: 'atx' },
       'no-missing-space-atx': true,
       'blanks-around-headings': true,
-      
+
       // Lists
       'blanks-around-lists': true,
       'list-indent': true,
@@ -25,35 +25,35 @@ export const PRESETS: Record<string, PresetConfig> = {
       'no-blanks-blockquote': true,
       'ul-style': { style: 'dash' },
       'ol-prefix': { style: 'ordered' },
-      
+
       // Code blocks
       'blanks-around-fences': true,
       'fenced-code-language': true,
       'code-fence-style': { style: 'backtick' },
-      
+
       // Links and references
       'no-bare-urls': true,
       'link-fragments': true,
       'reference-links-images': true,
-      
+
       // Line length
       'line-length': { line_length: 100, code_blocks: false, tables: false },
-      
+
       // Whitespace
       'no-trailing-spaces': true,
       'no-hard-tabs': true,
       'no-multiple-blanks': { maximum: 1 },
       'single-trailing-newline': true,
-      
+
       // Content
       'no-duplicate-heading': true,
-      'proper-names': { 
+      'proper-names': {
         names: ['JavaScript', 'TypeScript', 'GitHub', 'npm', 'Node.js'],
         code_blocks: false,
       },
     },
   },
-  
+
   standard: {
     description: 'Standard preset with sensible defaults',
     rules: {
@@ -61,33 +61,33 @@ export const PRESETS: Record<string, PresetConfig> = {
       'heading-increment': true,
       'heading-style': { style: 'atx' },
       'no-missing-space-atx': true,
-      
+
       // Lists
       'list-indent': true,
       'list-marker-space': true,
       'ul-style': { style: 'dash' },
-      
+
       // Code blocks
       'fenced-code-language': false,
       'code-fence-style': { style: 'consistent' },
-      
+
       // Links
       'no-bare-urls': true,
-      
+
       // Line length - disabled for flexibility
       'line-length': false,
-      
+
       // Whitespace
       'no-trailing-spaces': true,
       'no-hard-tabs': false, // Allow tabs in code blocks
       'no-multiple-blanks': { maximum: 2 },
       'single-trailing-newline': true,
-      
+
       // Content - more flexible
       'no-duplicate-heading': false,
     },
   },
-  
+
   relaxed: {
     description: 'Relaxed preset with minimal rules',
     rules: {
@@ -97,7 +97,7 @@ export const PRESETS: Record<string, PresetConfig> = {
       'list-marker-space': true,
       'no-trailing-spaces': true,
       'single-trailing-newline': true,
-      
+
       // Everything else disabled
       'line-length': false,
       'no-hard-tabs': false,
