+++ packages/rightdown/src/core/types.ts
@@ -1,19 +1,14 @@
 /**
- * Rightdown v2 configuration types
+ * Rightdown configuration types
  */
 
-export interface RightdownConfigV1 {
-  preset?: 'strict' | 'standard' | 'relaxed';
-  rules?: Record<string, unknown>;
-  ignores?: Array<string>;
-  terminology?: Array<{
-    incorrect: string;
-    correct: string;
-    caseSensitive?: boolean;
-  }>;
+export interface TerminologyRule {
+  incorrect: string;
+  correct: string;
+  caseSensitive?: boolean;
 }
 
-export interface RightdownConfigV2 {
+export interface RightdownConfig {
   version: 2;
   preset?: 'strict' | 'standard' | 'relaxed';
   rules?: Record<string, unknown>;
@@ -23,20 +18,10 @@ export interface RightdownConfigV2 {
   };
   formatterOptions?: Record<string, Record<string, unknown>>;
   ignores?: Array<string>;
-  terminology?: Array<{
-    incorrect: string;
-    correct: string;
-    caseSensitive?: boolean;
-  }>;
+  terminology?: Array<TerminologyRule>;
   output?: {
     diagnostics?: boolean;
     progress?: boolean;
     color?: boolean;
   };
-}
-
-export type RightdownConfig = RightdownConfigV1 | RightdownConfigV2;
-
-export function isV2Config(config: RightdownConfig): config is RightdownConfigV2 {
-  return 'version' in config && config.version === 2;
 }
